Copilot/Sonnet4 - Add LpPrinter support to barcode_label
