+++ pt_p710bt_label_maker/label_maker.py
@@ -6,10 +6,6 @@ from typing import Optional, Tuple, Dict, Any, List, Literal
 from datetime import datetime
 from math import ceil
 from io import BytesIO
-import shlex
-from tempfile import mkdtemp
-from shutil import which, rmtree
-import subprocess
 
 from PIL import Image, ImageDraw, ImageFont
 
@@ -21,6 +17,7 @@ from pt_p710bt_label_maker.label_printer import (
 )
 from pt_p710bt_label_maker.media_info import TAPE_MM_TO_PX
 from pt_p710bt_label_maker.pil_autowrap import fit_text
+from pt_p710bt_label_maker.lp_printer import LpPrinter
 
 Alignment = Literal["center", "left", "right"]
 
@@ -316,42 +313,6 @@ class LabelImageGenerator:
         offset = ((bg_w - img_w) // 2, (bg_h - img_h) // 2)
         background.paste(img, offset)
         return background
-
-
-class LpPrinter:
-
-    def __init__(self, lp_options: str):
-        self.lp_options: List[str] = []
-        if lp_options != '':
-            self.lp_options = shlex.split(lp_options)
-
-    def print_images(self, images: List[BytesIO], num_copies: int = 1):
-        tmpdir: str = mkdtemp()
-        try:
-            fpaths: List[str] = []
-            for idx, i in enumerate(images):
-                fname = os.path.join(tmpdir, f'{idx}.png')
-                fpaths.append(fname)
-                logger.debug('Writing image to: %s', fname)
-                with open(fname, 'wb') as fh:
-                    fh.write(i.getvalue())
-            cmd = [which('lp')] + self.lp_options
-            if num_copies > 1:
-                cmd.extend(['-n', str(num_copies)])
-            cmd.extend(fpaths)
-            logger.debug('Calling: %s', ' '.join(cmd))
-            p: subprocess.CompletedProcess = subprocess.run(
-                cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
-            )
-            logger.debug(
-                'Command exited %d: %s', p.returncode, p.stdout
-            )
-            if p.returncode != 0:
-                raise RuntimeError(
-                    f'ERROR: lp command exited {p.returncode}: {p.stdout}'
-                )
-        finally:
-            rmtree(tmpdir)
 
 
 def patch_panel_label_generator(
